Add multi-line strings
This commit introduces multi-line strings delimited by """.

Rationale:

These are primarly intended to make writing longer strings in Sail
attributes simpler, so one can do:
```
$[attr {
  some_key = """
    A multi-line
      string
    value
    """
}]
```
A second reason to add these is that the auto-formatter can make
a mess of multi-line strings made by repeatedly appending using `^`,
whereas these format in a straightforward way.

The rules are similar to C#.

- The opening """ must be follwed by a newline. This newline is ignored.

- The closing """ must be on it's own line, it can be preceeded by
  whitespace (tab or space) characters. This whitespace controls
  the indentation of the string. Each line after the opening """ line
  must be prefixed by exactly this sequence of whitespace characters.

- The final newline before the closing """ line is also ignored.

- Escape characters are handled on a per-line basis, as per the rules
  for ordinary Sail strings. This is a point of divergence from C#
  multi-line strings, as these are not also raw strings. This means
  that if a trailing newline is wanted, one can just do:
  ```
  """
  a multi-
  line string\n
  """
  ```

- Unlike regular strings " can appear unescaped.

- The trailing newline of any line within the multi-line string cannot
  be escaped, so
  ```
  """
  escape\
  newline
  """
  ```
  is not permitted.

If we consider the first example in this commit message, it would
produce the regular string:
```
"A multi-line\n  string\nvalue"
```