Improve attribute formatting
Allow long attributes to be formatted across multiple lines, and support
multi-line strings in attributes. Fix an issue with empty lines in
multi-line strings.

Also fix formatting of comments at the end of delimited constructs.
This is any construct where we have something like
```
<opener> X_0 <delim> ... <delim> X_n <opt_delim> <closer>
```
In the case where this something like
```sail
{
  X_0, // line comment
  ...,
  X_n, // last comment
}
```
then previously `// last comment` would fall out past the closing `}`.
This is slightly tricky because we need to distinguish the above case
from
```sail
X_n, } // last comment
```
where the closer of the delimited block `}` is on the same line before
the comment.