Unify bit and bitvector(1) (#1491)
* Unify bit and bitvector(1)

For mostly historical reasons Sail has had a separate type `bit`
which is distinct from bits(1). This was because (a long time ago),
`bitvector(N)` and `vector(N, bit)` were the same type. Therefore `bit`
could not be `vector(1, bit)` because that would be `vector(1, vector(1, bit))`
and so on. Now `bitvector` and `vector` are different, so there is
no reason to keep `bit` and `bitvector(1)` separate other than inertia
(except for one caveat, discussed below).

The downside of these types being different was that extracting a single
bit into a bitvector has to be written as `[v[0]]`, i.e. extract the `bit`
then put it back into a singleton bitvector. It also meant we had separate
literals `bitzero` and `bitone` for the bit type, which is a bit ugly.

The one caveat as mentioned for merging these two concepts is that it
makes it ever so slightly trickier to write a Sail backend that uses a
bitlist representation for bitvectors (in OCaml for example it is
quite natural to map `bivector(1)` to `bit list` and `bit` to `bit`).

The changes here are fairly substantial:

* The literal constructors L_zero and L_one are removed from the AST

* The V_bit value constructor is removed. V_vector is replaced with
  V_bitvector and V_vector.

* [x,y,z] is treated as a concatenation of length-1 bitvectors (when
  not a generic vector expression).

* For backwards compatability `bitzero` and `bitone` are aliases for
  `0b0` and `0b1` respectively.

* The bit type is removed, and replaced with a type synonym in the
  prelude.

* The CT_bit type is removed from the Jib IR. We could keep it for
  SystemVerilog generation, but it seems cleaner to just remove it too.

* Adjust pattern completeness checking behaviour when all patterns are impossible

* Fix interpreter semantics for bit and bits(1) unification

* Fix issues causing lem backend to loop forever

* Fix all Sail->C tests

* Get Sail->OCaml tests working again

* Fix some issues with large vector concat matches

* Fix SV tests

* Fix non-machine word lem tests

* Fix executable Lean tests

* Update Lean expected tests

* Remove void OCaml test

We no longer allow clauses with impossible constraints

* Remove now defunct rewrite

* Fixing lem machine word tests

* Remove separate eq_bit function from == overload

Leave it in stdlib, but equivalent to eq_bits

* Fix some Coq tests

This is a bit of hack, but mostly to see what it gets past CI
and what needs to change.

* Should be all Rocq tests passing

Again fix is a bit hacky

* Update tests

Mark one monomorphisation test as expected fail for now